requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.3.post0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.4
mysqlclient==2.1.1
packaging==21.3
prompt-toolkit==3.0.31
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.5.0
